Modification DAO

+++ Projet_prog_sys_PPC/Business/Actions.cs
@@ -10,7 +10,7 @@ namespace Projet_prog_sys_PPC.Business
         public string Name { get; set; }
         public string Description { get; set; }
         public People People { get; set; }
-        public float Time { get; set; }
+        public int Time { get; set; }
 
     }
 }
